requirements.txt
fastapi
uvicorn[standard]
jinja2
langchain
langchain-openai
langchain-huggingface
langchain-chroma
chromadb
sentence-transformers
python-dotenv
streamlit
# PDF & OCR
pymupdf
pytesseract
pillow